Check for nonstandard links

+++ boson/commands/public/rip.rb
@@ -164,17 +164,22 @@ module RipLib
     }
   end
 
-  # @options :delete=>:boolean
-  # Checks for broken symlinks
+  # @options :delete=>:boolean, :non_standard=>:boolean
+  # Checks for broken or nonstandard symlinks
   def rip_symlinks(*files)
     options = files[-1].is_a?(Hash) ? files.pop : {}
     files = files.empty? ?  Dir.glob(File.expand_path("~/.rip/*/**/*.rb")) :
       files.map {|e| File.directory?(e) ? Dir.glob(e+'/**/*') : e }.flatten
-    symlinks = files.select {|e| File.symlink?(e) }
+    symlinks = files.select {|e| File.symlink?(e) }.map {|e| [e, File.readlink(e)] }
     puts "Checking #{symlinks.size} symlinks"
-    broken = symlinks.map {|e| [e, File.readlink(e)] }.select {|k,v| !File.exists?(v) }
-    broken.each {|k,v| File.unlink(k) } if options[:delete]
-    broken
+    if options[:non_standard]
+      package_dir = File.expand_path("~/.rip/.packages")
+      symlinks.reject {|k,v| v[/^#{package_dir}/] }
+    else
+      broken = symlinks.select {|k,v| !File.exists?(v) }
+      broken.each {|k,v| File.unlink(k) } if options[:delete]
+      broken
+    end
   end
 
   # @options :verbose=>:boolean
